Use double computation

+++ src/mandelbrot_kernel.cuh
@@ -8,7 +8,7 @@ namespace FastMandelbrot
 
     void call_mandelbrot_kernel(
         cudaSurfaceObject_t surface, unsigned int width, unsigned int height,
-        float2 origin, float size, unsigned int step_count);
+        double2 origin, double size, unsigned int step_count);
 }
 
 #endif /*MANDELBROT_KERNEL_H_ */
